# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finanzen",
    "description": "Haushaltsbuch. Beträge sind durchgehend ganze Cent; jedes Geldfeld endet auf `Cents`. Kategorienamen, Typ-Labels und Kommentare sind Daten und bleiben deutsch. KitchenOwl ist ein getrenntes, paralleles Journal und wird nie mit den persönlichen Buchungen verrechnet.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Dieselbe Origin wie die Oberfläche"
    }
  ],
  "paths": {
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Alle Konten",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Nur für Administratoren",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Konto angelegt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "409": {
            "description": "Benutzername vergeben",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Registration is closed by default, so accounts are created here by an admin.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Angemeldet; Session als HttpOnly-Cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Benutzername oder Passwort falsch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Die angemeldete Person",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "put": {
        "tags": [
          "auth"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Passwort geändert, alle Sitzungen beendet"
          },
          "401": {
            "description": "Aktuelles Passwort falsch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Administratorkonto angelegt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "409": {
            "description": "Es existiert bereits ein Konto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "displayName",
          "password"
        ],
        "properties": {
          "displayName": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "PasswordRequest": {
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
      "SetupRequest": {
        "type": "object",
        "required": [
          "username",
          "displayName",
          "password"
        ],
        "properties": {
          "displayName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "displayName",
          "isAdmin"
        ],
        "properties": {
          "displayName": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}